refactor(module): split platform_kubernetes_kind.go into _core + _gke
Phase 0 precursor for cloud-SDK extraction. kindBackend/eksErrorBackend/
aksBackend (all SDK-free) move to platform_kubernetes_core.go with a core
init(); gkeBackend (the only SDK-bearing k8s backend) moves to
platform_kubernetes_gke.go with its own init(). Behavior-equivalent: same
five backend names registered. Isolates the lone SDK-bearing platform
file for a later clean deletion.
